refactor: update import paths and remove unused SVG file; add user profile header and screen size component
