+++ src/components/organisms/header/index.tsx
@@ -6,21 +6,29 @@ import React, { useState } from 'react';
 import { Menu, X } from 'lucide-react';
 import Link from 'next/link';
 
-const Header: React.FC = () => {
+export function Header() {
   const [menuOpen, setMenuOpen] = useState(false);
 
   const toggleMenu = () => {
     setMenuOpen(!menuOpen);
   };
 
+  const navTitles = [
+    { title: 'Notícias', href: 'https://recoleta-news.vercel.app/' },
+    { title: 'Sobre nós', href: '#' },
+    { title: 'Nosso impacto', href: '#' },
+    { title: 'Parceiros', href: '#' },
+  ];
+
   return (
-    <header className="grid grid-cols-6 items-center px-6 relative z-10 pt-10">
+    <header className="grid grid-cols-6 items-start px-6 relative z-10 pt-10">
       <span className="flex flex-row items-center gap-4">
         <Image
           src="/images/logo_recoleta.svg"
           alt="Logo"
           width={60}
           height={60}
+          quality={80}
         />
         <h1
           className="xl:text-3xl text-2xl font-bold text-[#F6F6F6]"
@@ -30,23 +38,16 @@ const Header: React.FC = () => {
         </h1>
       </span>
 
-      <nav className="hidden md:block col-span-4 self-center justify-center">
+      <nav className="hidden md:block col-span-4 justify-center pt-3">
         <ul
           className="flex flex-row font-bold text-[#F6F6F6] justify-center 2xl:space-x-24 xl:space-x-12 md:gap-x-2 space-x-5 xl:text-xl text-base"
           style={{ textShadow: '1px 2px 4px rgba(0, 0, 0, 0.7)' }}
         >
-          <li>
-            <Link href="https://recoleta-news.vercel.app/">Notícias</Link>
-          </li>
-          <li>
-            <a href="#">Sobre nós</a>
-          </li>
-          <li>
-            <a href="#">Nosso impacto</a>
-          </li>
-          <li>
-            <a href="#">Parceiros</a>
-          </li>
+          {navTitles.map(nav => (
+            <li key={nav.title}>
+              <a href={nav.href}>{nav.title}</a>
+            </li>
+          ))}
         </ul>
       </nav>
 
@@ -56,35 +57,26 @@ const Header: React.FC = () => {
         </button>
       </span>
 
-      <span className="flex justify-end">
-       <Button
+      <span className="flex justify-end pt-3">
+        <Button
           variant="default"
           className="bg-white text-green-800 border border-green-800 font-bold"
         >
-          <a href="/login">Acessar </a>
+          <Link href="/login">Acessar</Link>
         </Button>
       </span>
 
       {menuOpen && (
         <nav className="absolute top-full left-0 w-full bg-white p-4 md:hidden">
           <ul className="flex flex-col font-extrabold text-black gap-4 text-xl">
-            <li>
-              <a href="#">Notícias</a>
-            </li>
-            <li>
-              <a href="#">Sobre nós</a>
-            </li>
-            <li>
-              <a href="#">Nosso impacto</a>
-            </li>
-            <li>
-              <a href="#">Parceiros</a>
-            </li>
+            {navTitles.map(nav => (
+              <li key={nav.title}>
+                <a href={nav.href}>{nav.title}</a>
+              </li>
+            ))}
           </ul>
         </nav>
       )}
     </header>
   );
-};
-
-export default Header;
+}
