+++ src/components/organisms/sidebar/index.tsx
@@ -145,7 +145,7 @@ export function SidebarComponent({
         <SidebarMenu>
           <SidebarMenuItem>
             <SidebarMenuButton size="lg" asChild>
-              <a href="#">
+              <a href="/dashboard">
                 <div className="flex aspect-square size-8 items-center justify-center rounded-lg bg-sidebar-primary text-sidebar-primary-foreground">
                   <Image
                     alt="logo-recoleta"
